pyproject.toml
[tool.poetry]
name = "agentic-patterns"
version = "0.0.1"
description = "Agentic Patterns"
authors = ["Serg Masis"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
crewai = ">=0.83.0"
crewai-tools = ">=0.14.0"
exa-py = "*"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
coverage = {extras = ["toml"], version = "7.6.*"}
black = {version = "24.8.*", allow-prereleases = true}
pytest = "8.3.*"
pytest-html = "4.1.*"
pytest-cov = "5.0.*"
pylint = "3.2.*"
python-decouple = "3.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
